rename linera_service/src/client.rs into cli_wrappers.rs
