+++ linera-service/src/linera.rs
@@ -26,7 +26,7 @@ use linera_execution::{
 use linera_rpc::node_provider::{NodeOptions, NodeProvider};
 use linera_service::{
     chain_listener::{self, ChainListenerConfig},
-    client::{Database, LocalNetwork, Network},
+    cli_wrappers::{Database, LocalNetwork, Network},
     config::{CommitteeConfig, Export, GenesisConfig, Import, UserChain, WalletState},
     node_service::NodeService,
     project::{self, Project},
